Cleanup main.rs
