+++ src-tauri/src/frontend_types.rs
@@ -7,3 +7,37 @@ pub struct ChampionInfo {
     pub pick_rate: String,
     pub ban_rate: String,
 }
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RuneImages {
+    pub primary_runes: PrimaryTree,
+    pub secondary_runes: SecondaryTree,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PrimaryTree {
+    pub slot_one: Vec<Active>,
+    pub slot_two: Vec<Active>,
+    pub slot_three: Vec<Active>,
+    pub slot_four: Vec<Active>,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SecondaryTree {
+    pub slot_one: Vec<Active>,
+    pub slot_two: Vec<Active>,
+    pub slot_three: Vec<Active>,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Active {
+    pub name: String,
+    pub image: String,
+    pub local_image: String,
+    pub active: bool,
+    pub id: i64,
+}
